Move createAGlobalLookupDictionary test to own file
